refactoring main
